requirementschat.txt

django-widget-tweaks==1.4.1
djangorestframework==3.9.1
pytz==2017.3
python-memcached